Домашняя работа 1. Задание 3. Исправления

+++ src/Main.java
@@ -78,7 +78,7 @@ public class Main {
         if (deliveryDistance <= 20) {
             return 1;
         } else {
-            return (int) Math.round (deliveryDistance / 40) + 1;
+            return (int) Math.round ((double) (deliveryDistance / 40)) + 1;
         }
     }
 }
